third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-v2-events.html

<!DOCTYPE html>
<html>
    <head>
        <title>MediaKeys</title>
        <script src=../video-test.js></script>
        <script>
            function stringToUint8Array(str)
            {
                var arr=[];
                for(var i=0,j=str.length;i<j;++i)
                  arr[i]=str.charCodeAt(i);
                return new Uint8Array(arr);
            }

            var mediaKeys;
            var mediaKeySession;
            var initData = stringToUint8Array('mock');
            var invalidKey = stringToUint8Array('invalid');
            var validKey = stringToUint8Array(
                '{"keys":[{"kty":"oct","kid":"AAECAw","k":"MTIzNDU2Nzg5MGFiY2RlZg"}]}');

            function runTest() 
            {
                run('mediaKeys = new MediaKeys("org.w3.clearkey")');
                run('mediaKeySession = mediaKeys.createSession("video/webm", initData)');
                waitForEvent('message', processMessage, false, false, mediaKeySession);
                waitForEvent('error', processError, false, false, mediaKeySession);
                waitForEvent('ready', processReady, false, false, mediaKeySession);
                waitForEvent('close', null, true, false, mediaKeySession);
            }

            function processMessage(event)
            {
                run('mediaKeySession.update(invalidKey)');
            }

            function processError(event)
            {
                run('mediaKeySession.update(validKey)');
            }

            function processReady(event)
            {
                run('mediaKeySession.release()');
            }
        </script>
    </head>
    <body onload="runTest()">
        <p>This tests the events created by MediaKeys.</p>
    </body>
</html>
