third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-v2-events-expected.txt

This tests the events created by MediaKeys.

RUN(mediaKeys = new MediaKeys("org.w3.clearkey"))
RUN(mediaKeySession = mediaKeys.createSession("video/webm", initData))
EVENT(message)
RUN(mediaKeySession.update(invalidKey))
EVENT(error)
RUN(mediaKeySession.update(validKey))
EVENT(ready)
RUN(mediaKeySession.release())
EVENT(close)
END OF TEST
